move Events Header + search into container so that it appears while loading
